게시글 / 댓글 관리 기능 추가

+++ History.MobileClient/ViewModels/CommentViewModel.cs
@@ -3,6 +3,7 @@ using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
 using CommunityToolkit.Mvvm.Messaging.Messages;
 using History.Commons.Api.Comment;
+using History.Commons.Api.Moderation;
 using History.Commons.DataTypes.ResponseDtos;
 using History.Commons.Enums;
 using History.MobileClient.DataTypes;
@@ -102,11 +103,28 @@ public partial class CommentViewModel : ObservableObject
 
     public async Task DeleteAsync()
     {
-        var result = await App.Page.DisplayAlert("댓글 삭제", "정말로 댓글을 삭제하시겠습니까?", Constants.PromptOk, Constants.PromptCancel);
-        if (!result) return;
+        if(Comment.User.UserId != Shared.UserId)
+        {
+            if (Shared.MyRank < Rank.Moderator)
+            {
+                await App.Page.DisplayAlert("권한 부족", "댓글을 삭제할 권한이 없습니다.", Constants.PromptOk);
+                return;
+            }
 
-        var commentResult = await App.ExecuteRequestAsync(new DeleteComment(Comment.Id));
-        if (commentResult.IsSuccess) WeakReferenceMessenger.Default.Send(new ValueDeletedMessage<CommentResponseDto>(Comment));
+            var reason = await App.Page.DisplayPromptAsync("댓글 삭제", "댓글을 삭제하는 이유를 입력해주세요.", "삭제", "취소", "삭제 사유");
+            if (string.IsNullOrWhiteSpace(reason)) return;
+
+            var deleteResult = await App.ExecuteRequestAsync(new RestrictionDeleteComment(Comment.Id, reason));
+            if (deleteResult.IsSuccess) WeakReferenceMessenger.Default.Send(new ValueDeletedMessage<CommentResponseDto>(Comment));
+        }
+        else
+        {
+            var result = await App.Page.DisplayAlert("댓글 삭제", "정말로 댓글을 삭제하시겠습니까?", Constants.PromptOk, Constants.PromptCancel);
+            if (!result) return;
+
+            var deleteResult = await App.ExecuteRequestAsync(new DeleteComment(Comment.Id));
+            if (deleteResult.IsSuccess) WeakReferenceMessenger.Default.Send(new ValueDeletedMessage<CommentResponseDto>(Comment));
+        }
     }
 
     [RelayCommand]
